requirements.txt
annotated-types==0.6.0
anyio==4.3.0
behave==1.2.6
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
fastapi==0.110.0
h11==0.14.0
idna==3.6
parse==1.20.1
parse-type==0.6.2
pydantic==2.6.3
pydantic_core==2.16.3
python-dotenv==1.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.27.1
